requirements.txt
python-dotenv~=1.0.1
google-generativeai~=0.5.4
Werkzeug~=3.0.1
gunicorn~=23.0.0
bcrypt~=4.1.2
boto3~=1.34.55
pandas~=2.2.1
openpyxl~=3.1.2
requests~=2.31.0
Flask~=3.0.2
flask_cors~=5.0.0
langchain~=0.1.12
langchain-google-genai~=1.0.4
chromadb~=0.5.5
langchain-openai~=0.1.7
flashrank~=0.2.9
onnxruntime~=1.19.2
